+++ fixtures/base/collect-data.js
@@ -4,6 +4,8 @@ import { fileURLToPath } from 'url';
 import { createGitReader, parseTree } from '../../lib/index.js';
 
 const repo = await createGitReader(fileURLToPath(new URL('_git', import.meta.url)));
+const defaultBranch = await repo.defaultBranch();
+const currentBranch = await repo.currentBranch();
 const commits = await repo.log();
 const stat = await repo.stat();
 const objects = {};
@@ -38,6 +40,8 @@ fs.writeFileSync(
     fileURLToPath(new URL('data.json', import.meta.url)),
     JSON.stringify(
         {
+            defaultBranch,
+            currentBranch,
             commits,
             filesLists,
             filesDelta,