+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Added `repo.describeRef(ref)` method, which returns an information object about the reference
+
 ## 0.1.3 (2023-10-13)
 
 - Fixed size computation of on-disk reverse index (`.rev` files)